requirements.txt
transformers==4.36.1
accelerate==0.24.0
bitsandbytes==0.41.3.post2
datasets==2.15.0
peft==0.7.0
trl==0.7.4
sentence-transformers==2.2.2
hydra-core
pynvml
sentencepiece
einops
numpy<2
pandas
scikit-learn
protobuf